Authentication and Authorization added with exception handling

+++ src/main/java/com/yash/project/HotelManagementApp/service/HotelServiceImpl.java
@@ -1,16 +1,19 @@
 package com.yash.project.HotelManagementApp.service;
 
 import com.yash.project.HotelManagementApp.Exception.ResourceNotFoundException;
+import com.yash.project.HotelManagementApp.Exception.UnAuthorisedException;
 import com.yash.project.HotelManagementApp.dto.HotelDto;
 import com.yash.project.HotelManagementApp.dto.HotelInfoDto;
 import com.yash.project.HotelManagementApp.dto.RoomDto;
 import com.yash.project.HotelManagementApp.entity.Hotel;
 import com.yash.project.HotelManagementApp.entity.Room;
+import com.yash.project.HotelManagementApp.entity.User;
 import com.yash.project.HotelManagementApp.repository.HotelRepository;
 import com.yash.project.HotelManagementApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -31,6 +34,11 @@ public class HotelServiceImpl implements HotelService{
         log.info("Creating a new Hotel with name:{}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto,Hotel.class);
         hotel.setActive(false);
+
+        //checking authorization
+        User user = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel=hotelRepository.save(hotel);
         log.info("Created a hotel with ID:{}",hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class);
@@ -42,6 +50,13 @@ public class HotelServiceImpl implements HotelService{
         log.info("Getting the hotel with id:{}",id);
         Hotel hotel = hotelRepository.findById(id).
                 orElseThrow(()-> new ResourceNotFoundException("Hotel not found with id:"+id));
+
+        //checking authorization
+        User user = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -50,6 +65,12 @@ public class HotelServiceImpl implements HotelService{
         log.info("Updating the hotel with id:{}",id);
         Hotel hotel = hotelRepository.findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with id:"+id));
+
+        //checking authorization
+        User user = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
 
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
@@ -64,7 +85,11 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id:" + id));
 
-
+        //checking authorization
+        User user = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
 
         for(Room room: hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
@@ -82,6 +107,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository.findById(HotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id:" + HotelId));
         hotel.setActive(true);
+
+        //checking authorization
+        User user = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+HotelId);
+        }
 
         //assuming only do it once
         for(Room room : hotel.getRooms()){
