+++ src/main/java/com/yash/project/HotelManagementApp/service/RoomServiceImpl.java
@@ -1,14 +1,17 @@
 package com.yash.project.HotelManagementApp.service;
 
 import com.yash.project.HotelManagementApp.Exception.ResourceNotFoundException;
+import com.yash.project.HotelManagementApp.Exception.UnAuthorisedException;
 import com.yash.project.HotelManagementApp.dto.RoomDto;
 import com.yash.project.HotelManagementApp.entity.Hotel;
 import com.yash.project.HotelManagementApp.entity.Room;
+import com.yash.project.HotelManagementApp.entity.User;
 import com.yash.project.HotelManagementApp.repository.HotelRepository;
 import com.yash.project.HotelManagementApp.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,6 +33,13 @@ public class RoomServiceImpl implements RoomService{
         log.info("Getting the hotel with id:{}",hotelId);
         Hotel hotel = hotelRepository.findById(hotelId).
                 orElseThrow(()-> new ResourceNotFoundException("Hotel not found with id:"+hotelId));
+
+        //checking authorization
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         log.info("Creating a new room");
         Room room = modelMapper.map(roomDto,Room.class);
         room.setHotel(hotel);
@@ -55,6 +65,8 @@ public class RoomServiceImpl implements RoomService{
     @Override
     public RoomDto getRoomById(Long roomId) {
         log.info("Getting the room with ID: {}", roomId);
+
+
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
@@ -68,6 +80,13 @@ public class RoomServiceImpl implements RoomService{
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
+
+        //checking authorization
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorisedException("This user does not own this room with id: "+roomId);
+        }
+
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
